created layout
