+++ flatties-frontend/src/pages/home/HomePage.tsx
@@ -1,13 +1,11 @@
-import React from 'react';
-import NavBar from '../../components/navBar/NavBar';
+import React from "react";
+import NavBar from "../../components/layout/navBar/NavBar";
 
 function HomePage() {
   return (
     <div>
-      <NavBar />
       <h1>Hello, world!</h1>
       <h1>Home page</h1>
-
     </div>
   );
 }
