+++ flatties-frontend/src/pages/listing/ListingPage.tsx
@@ -1,14 +1,12 @@
-
-import React from 'react';
-import NavBar from '../../components/navBar/NavBar';
+import React from "react";
+import NavBar from "../../components/layout/navBar/NavBar";
 
 function ListingPage() {
   return (
     <div>
-      <NavBar />
       <h1>Listing Page</h1>
     </div>
   );
-};
+}
 
 export default ListingPage;
