+++ flatties-frontend/src/pages/profile/ProfilePage.tsx
@@ -1,14 +1,12 @@
-
-import React from 'react';
-import NavBar from '../../components/navBar/NavBar';
+import React from "react";
+import NavBar from "../../components/layout/navBar/NavBar";
 
 function ProfilePage() {
   return (
     <div>
-      <NavBar />
       <h1>Profile Page</h1>
     </div>
   );
-};
+}
 
 export default ProfilePage;
