+++ flatties-frontend/src/pages/register/RegisterPage.tsx
@@ -1,14 +1,12 @@
-
-import React from 'react';
-import NavBar from '../../components/navBar/NavBar';
+import React from "react";
+import NavBar from "../../components/layout/navBar/NavBar";
 
 function RegisterPage() {
   return (
     <div>
-      <NavBar />
       <h1>Register Page</h1>
     </div>
   );
-};
+}
 
 export default RegisterPage;
